refactor: extrae header a su propio componeente
