+++ src/ui/views/_components/Header/_components/LanguageSelector/LanguageSelector.module.scss
@@ -1,0 +1,28 @@
+.languageButtonsContainer {
+    display: flex;
+    margin-right: 20px;
+    align-self: flex-end;
+    margin-left: auto;
+    margin-bottom: 10px;
+  }
+
+
+.languageButton {
+    padding: 8px 12px;
+    margin: 0 5px;
+    background-color: white;
+    cursor: pointer;
+    border-radius: 10px;
+    border: 1px solid #e0e0e0;
+    box-shadow: 0 2px 5px rgba(0,0,0,0.1);
+    justify-self: flex-end;
+    
+    &:hover {
+      background-color: #f5f5f5;
+    }
+    
+    &.active {
+    font-weight: bold;
+    background-color: #e0e0e0;
+  }
+}
