+++ src/ui/views/_components/Header/_components/LanguageSelector/index.ts
@@ -1,0 +1,1 @@
+export { LanguageSelector } from "./LanguageSelector";
